fix: centralise signals-related hogql
