+++ products/signals/backend/temporal/__init__.py
@@ -18,13 +18,10 @@ from products.signals.backend.temporal.grouping import (
     TeamSignalGroupingWorkflow,
     assign_and_emit_signal_activity,
     fetch_report_contexts_activity,
-    fetch_signal_type_examples_activity,
     generate_search_queries_activity,
     get_embedding_activity,
     match_signal_to_report_activity,
-    run_signal_semantic_search_activity,
     verify_match_specificity_activity,
-    wait_for_signal_in_clickhouse_activity,
 )
 from products.signals.backend.temporal.grouping_v2 import TeamSignalGroupingV2Workflow, read_signals_from_s3_activity
 from products.signals.backend.temporal.reingestion import (
@@ -35,9 +32,14 @@ from products.signals.backend.temporal.reingestion import (
 )
 from products.signals.backend.temporal.report_safety_judge import report_safety_judge_activity
 from products.signals.backend.temporal.safety_filter import safety_filter_activity
+from products.signals.backend.temporal.signal_queries import (
+    fetch_signal_type_examples_activity,
+    fetch_signals_for_report_activity,
+    run_signal_semantic_search_activity,
+    wait_for_signal_in_clickhouse_activity,
+)
 from products.signals.backend.temporal.summary import (
     SignalReportSummaryWorkflow,
-    fetch_signals_for_report_activity,
     mark_report_failed_activity,
     mark_report_in_progress_activity,
     mark_report_pending_input_activity,
